Extract common deployment step descriptor code into helper class
