+++ src/test/java/com/atlassian/bitbucket/jenkins/internal/deployments/DeploymentStepImplTest.java
@@ -2,14 +2,12 @@ package com.atlassian.bitbucket.jenkins.internal.deployments;
 
 import com.atlassian.bitbucket.jenkins.internal.model.deployment.BitbucketDeploymentEnvironment;
 import com.atlassian.bitbucket.jenkins.internal.model.deployment.BitbucketDeploymentEnvironmentType;
-import com.atlassian.bitbucket.jenkins.internal.provider.JenkinsProvider;
 import hudson.model.FreeStyleProject;
 import hudson.model.Item;
 import hudson.model.Run;
 import hudson.model.TaskListener;
 import hudson.util.FormValidation;
 import hudson.util.ListBoxModel;
-import jenkins.model.Jenkins;
 import org.jenkinsci.plugins.workflow.steps.StepContext;
 import org.jenkinsci.plugins.workflow.steps.StepExecution;
 import org.junit.Before;
@@ -24,7 +22,6 @@ import java.io.PrintStream;
 import java.net.URI;
 import java.util.logging.Logger;
 
-import static com.atlassian.bitbucket.jenkins.internal.deployments.DeploymentStepImpl.DescriptorImpl.FORM_VALIDATION_OK;
 import static java.lang.String.format;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.*;
@@ -36,7 +33,9 @@ import static org.mockito.Mockito.*;
 public class DeploymentStepImplTest {
 
     private static final Logger log = Logger.getLogger(DeploymentStepImplTest.class.getName());
-
+    private static final FormValidation FORM_VALIDATION_OK = FormValidation.ok();
+    @Mock
+    DeploymentStepDescriptorHelper descriptorHelper;
     @Mock
     private Run<FreeStyleProject, ?> run;
     @InjectMocks
@@ -47,11 +46,6 @@ public class DeploymentStepImplTest {
     private TaskListener taskListener;
     @Mock
     private PrintStream printStream;
-    @Mock
-    private JenkinsProvider jenkinsProvider;
-    @Mock
-    private Jenkins jenkins;
-
     private DeploymentStepImpl deploymentStep;
 
     @Before
@@ -60,151 +54,54 @@ public class DeploymentStepImplTest {
         when(context.get(Run.class)).thenReturn(run);
         when(context.get(TaskListener.class)).thenReturn(taskListener);
         when(taskListener.getLogger()).thenReturn(printStream);
-        when(jenkinsProvider.get()).thenReturn(jenkins);
-    }
-
-    @Test
-    public void testDescriptorDoCheckEnvironmentNameContextNull() {
-        String environmentName = "my env";
-
-        FormValidation formValidation = descriptor.doCheckEnvironmentName(null, environmentName);
-
-        verify(jenkins).checkPermission(Jenkins.ADMINISTER);
-        assertThat(formValidation, equalTo(FORM_VALIDATION_OK));
     }
 
     @Test
     public void testDescriptorDoCheckEnvironmentName() {
         Item context = mock(Item.class);
         String environmentName = "my env";
+        when(descriptorHelper.doCheckEnvironmentName(context, environmentName)).thenReturn(FORM_VALIDATION_OK);
 
         FormValidation formValidation = descriptor.doCheckEnvironmentName(context, environmentName);
 
-        verify(context).checkPermission(Item.EXTENDED_READ);
-        verifyZeroInteractions(jenkins);
-        assertThat(formValidation, equalTo(FORM_VALIDATION_OK));
-    }
-
-    @Test
-    public void testDescriptorDoCheckEnvironmentNameBlank() {
-        Item context = mock(Item.class);
-
-        FormValidation formValidation = descriptor.doCheckEnvironmentName(context, " ");
-
-        verify(context).checkPermission(Item.EXTENDED_READ);
-        verifyZeroInteractions(jenkins);
-        assertThat(formValidation.getMessage(), equalTo("The environment name is required."));
-    }
-
-    @Test
-    public void testDescriptorDoCheckEnvironmentTypeContextNull() {
-        FormValidation formValidation = descriptor.doCheckEnvironmentType(null, "Production");
-
-        verify(jenkins).checkPermission(Jenkins.ADMINISTER);
-        assertThat(formValidation, equalTo(FORM_VALIDATION_OK));
-    }
-
-    @Test
-    public void testDescriptorDoCheckEnvironmentTypeBlank() {
-        Item context = mock(Item.class);
-
-        FormValidation formValidation = descriptor.doCheckEnvironmentType(context, " ");
-
-        verify(context).checkPermission(Item.EXTENDED_READ);
-        verifyZeroInteractions(jenkins);
+        verify(descriptorHelper).doCheckEnvironmentName(context, environmentName);
         assertThat(formValidation, equalTo(FORM_VALIDATION_OK));
     }
 
     @Test
     public void testDescriptorDoCheckEnvironmentType() {
         Item context = mock(Item.class);
+        String environmentType = "PRODUCTION";
+        when(descriptorHelper.doCheckEnvironmentType(context, environmentType)).thenReturn(FORM_VALIDATION_OK);
 
-        FormValidation formValidation = descriptor.doCheckEnvironmentType(context, "PRODUCTION");
+        FormValidation formValidation = descriptor.doCheckEnvironmentType(context, environmentType);
 
-        verify(context).checkPermission(Item.EXTENDED_READ);
-        verifyZeroInteractions(jenkins);
+        verify(descriptorHelper).doCheckEnvironmentType(context, environmentType);
         assertThat(formValidation, equalTo(FORM_VALIDATION_OK));
-    }
-
-    @Test
-    public void testDescriptorDoCheckEnvironmentTypeBadEnvironmentType() {
-        Item context = mock(Item.class);
-
-        FormValidation formValidation = descriptor.doCheckEnvironmentType(context, "not an environment type");
-
-        verify(context).checkPermission(Item.EXTENDED_READ);
-        verifyZeroInteractions(jenkins);
-        assertThat(formValidation.getMessage(), equalTo("The environment type should be one of DEVELOPMENT, PRODUCTION, STAGING, TESTING."));
-    }
-
-    @Test
-    public void testDescriptorDoCheckEnvironmentUrlContextNull() {
-        FormValidation formValidation = descriptor.doCheckEnvironmentUrl(null, "http://my-env");
-
-        verify(jenkins).checkPermission(Jenkins.ADMINISTER);
-        assertThat(formValidation, equalTo(FORM_VALIDATION_OK));
-    }
-
-    @Test
-    public void testDescriptorDoCheckEnvironmentUrlBlank() {
-        Item context = mock(Item.class);
-
-        FormValidation formValidation = descriptor.doCheckEnvironmentUrl(context, " ");
-
-        verify(context).checkPermission(Item.EXTENDED_READ);
-        verifyZeroInteractions(jenkins);
-        assertThat(formValidation, equalTo(FORM_VALIDATION_OK));
-    }
-
-    @Test
-    public void testDescriptorDoCheckEnvironmentUrlInvalid() {
-        Item context = mock(Item.class);
-
-        FormValidation formValidation = descriptor.doCheckEnvironmentUrl(context, "not a url");
-
-        verify(context).checkPermission(Item.EXTENDED_READ);
-        verifyZeroInteractions(jenkins);
-        assertThat(formValidation.getMessage(), equalTo("The environment URL must be a valid URL."));
-    }
-
-    @Test
-    public void testDescriptorDoCheckEnvironmentUrlNotAbsolute() {
-        Item context = mock(Item.class);
-
-        FormValidation formValidation = descriptor.doCheckEnvironmentUrl(context, "/relative/url");
-
-        verify(context).checkPermission(Item.EXTENDED_READ);
-        verifyZeroInteractions(jenkins);
-        assertThat(formValidation.getMessage(), equalTo("The deployment URI must be absolute."));
-    }
-
-    @Test
-    public void testDoFillEnvironmentTypeItems() {
-        ListBoxModel options = descriptor.doFillEnvironmentTypeItems(null);
-
-        verify(jenkins).checkPermission(Jenkins.ADMINISTER);
-        assertThat(options, hasSize(5));
-        assertThat(options.get(0).name, equalTo("- none -"));
-        assertThat(options.get(0).value, equalTo(""));
-        assertThat(options.get(1).name, equalTo("Production"));
-        assertThat(options.get(1).value, equalTo("PRODUCTION"));
-        assertThat(options.get(2).name, equalTo("Staging"));
-        assertThat(options.get(2).value, equalTo("STAGING"));
-        assertThat(options.get(3).name, equalTo("Testing"));
-        assertThat(options.get(3).value, equalTo("TESTING"));
-        assertThat(options.get(4).name, equalTo("Development"));
-        assertThat(options.get(4).value, equalTo("DEVELOPMENT"));
     }
 
     @Test
     public void testDescriptorDoCheckEnvironmentUrl() {
         Item context = mock(Item.class);
+        String environmentUrl = "http://my-env";
+        when(descriptorHelper.doCheckEnvironmentUrl(context, environmentUrl)).thenReturn(FORM_VALIDATION_OK);
 
-        FormValidation formValidation = descriptor.doCheckEnvironmentUrl(context, "http://my-env");
+        FormValidation formValidation = descriptor.doCheckEnvironmentUrl(context, environmentUrl);
 
-        verify(context).checkPermission(Item.EXTENDED_READ);
-        verifyZeroInteractions(jenkins);
+        verify(descriptorHelper).doCheckEnvironmentUrl(context, environmentUrl);
         assertThat(formValidation, equalTo(FORM_VALIDATION_OK));
+    }
+
+    @Test
+    public void testDoFillEnvironmentTypeItems() {
+        Item context = mock(Item.class);
+        ListBoxModel listBoxModel = mock(ListBoxModel.class);
+        when(descriptorHelper.doFillEnvironmentTypeItems(context)).thenReturn(listBoxModel);
+
+        ListBoxModel options = descriptor.doFillEnvironmentTypeItems(context);
+
+        verify(descriptorHelper).doFillEnvironmentTypeItems(context);
+        assertThat(options, equalTo(listBoxModel));
     }
 
     @Test
